+++ CHANGELOG.md
@@ -10,6 +10,7 @@ GAMSPy 1.9.0
   - Make GAMSPy symbols non-iterable.
   - Inherit output argument from the container in solve function if the output argument is not specified.
   - Rename ModelInstanceOptions to FreezeOptions and add a warning for the usage of ModelInstanceOptions.
+  - Fix sense=feasibility bug of frozen models. 
 
 GAMSPy 1.8.0
 ------------